# src/main/java/com/shamith/bookmate/auth/modal/AuthenticationResponse.java
package com.shamith.bookmate.auth.modal;

import lombok.Builder;
import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
@Builder
public class AuthenticationResponse {
    private String token;
}